Migrate common modules to pt2itp-lib crate
Having the neon bindings and pure rust code in the same crate causes
issues with using standard rust tooling for profiling, benchmarking,
etc.. due to the neon bindings producing a C library instead of an rlib.
Separating the native components into two crates allows the neon
bindings to live separately from the rust implementations, enabling the
use of standard rust development tools.

As a first step we introduce the pt2itp-lib crate which will contain the
rust implementations (neon bindings continue to live in pt2itp) and
migrate over the following common modules:

- pg
- text
- types

Future work will migrate the core pt2itp actions to pt2itp-lib.
