+++ .circleci/config.yml
@@ -43,7 +43,7 @@ jobs:
                 command: yarn build && yarn install
             - run:
                 name: "Cargo Test"
-                command: cd native/ && cargo test --release
+                command: cd native/ && cargo test --release --workspace
             - run:
                 name: "Cargo Format"
                 command: rustup component add rustfmt && cd native/ && cargo fmt -- --check
