+++ native/pt2itp-lib/.gitignore
@@ -1,0 +1,2 @@
+target
+
