+++ native/pt2itp-lib/Cargo.toml
@@ -1,0 +1,31 @@
+[package]
+name = "pt2itp-lib"
+version = "0.1.0"
+authors = ["Christopher Grim <chris.grim@mapbox.com>"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+fancy-regex = "0.1.0"
+geocoder-abbreviations = { git = "https://github.com/mapbox/geocoder-abbreviations", rev = "master" }
+lazy_static = "1.3.0"
+memchr = "2.0.2"
+postgis = "0.6.0"
+regex = "1.1"
+serde = "1.0"
+serde_derive = "1.0"
+serde_json = "1.0"
+unicode-segmentation = "1.3.0"
+
+[dependencies.geojson]
+version = "0.16.0"
+features = [
+    "geo-types"
+]
+
+[dependencies.postgres]
+version = "0.15.2"
+features = [
+    "with-serde_json"
+]
