+++ native/pt2itp-lib/src/pg/intersections.rs
