+++ native/pt2itp-lib/src/types/network.rs
